libtrap - added new file for common functions used in socket based interfaces
