+++ libtrap/src/ifc_tcpip.c
@@ -67,6 +67,7 @@
 #include "trap_error.h"
 #include "ifc_tcpip.h"
 #include "ifc_tcpip_internal.h"
+#include "ifc_socket_common.h"
 
 /**
  * \addtogroup trap_ifc TRAP communication module interface
@@ -957,55 +958,6 @@ static int client_socket_connect(void *priv, const char *dest_addr, const char *
  * @{
  */
 
-
-/**
- * \brief Array containing constants used for operations with bit arrays.
- */
-static uint64_t mask[64] = {
-        1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096, 8192, 16384, 32768, 65536, 131072, 262144, 524288,
-        1048576, 2097152, 4194304, 8388608, 16777216, 33554432, 67108864, 134217728, 268435456, 536870912, 1073741824,
-        2147483648, 4294967296, 8589934592, 17179869184, 34359738368, 68719476736, 137438953472, 274877906944,
-        549755813888, 1099511627776, 2199023255552, 4398046511104, 8796093022208, 17592186044416, 35184372088832,
-        70368744177664, 140737488355328, 281474976710656, 562949953421312, 1125899906842624, 2251799813685248,
-        4503599627370496, 9007199254740992, 18014398509481984, 36028797018963968, 72057594037927936, 144115188075855872,
-        288230376151711744, 576460752303423488, 1152921504606846976, 2305843009213693952, 4611686018427387904, 9223372036854775808ULL
-};
-
-/**
- * \brief Set i-th element (one bit) of 'bits' to 1.
- *
- * \param[in] bits Pointer to the bit array.
- * \param[in] i Target element's index in the 'bits' array.
- */
-static inline void set_index(uint64_t *bits, int i)
-{
-   *bits |= mask[i];
-}
-
-/**
- * \brief Set i-th element (one bit) of 'bits' to 0.
- *
- * \param[in] bits Pointer to the bit array.
- * \param[in] i Target element's index in the 'bits' array.
- */
-static inline void del_index(uint64_t *bits, int i)
-{
-   *bits &= (0xffffffffffffffff - mask[i]);
-}
-
-/**
- * \brief Return value of i-th element (one bit) in the 'bits' array.
- *
- * \param[in] bits Pointer to the bit array.
- * \param[in] i Target element's index in the 'bits' array.
- *
- * \return Value of i-th element (one bit) in the 'bits' array.
- */
-static inline uint64_t check_index(uint64_t bits, int i)
-{
-   return (bits & mask[i]);
-}
-
 /**
  * \brief This function is called when a client was/is being disconnected.
  *
@@ -1421,21 +1373,6 @@ static void *sending_thread_func(void *priv)
          }
       }
    }
-}
-
-/**
- * \brief Write data into buffer
- *
- * \param[in] buffer Pointer to the buffer.
- * \param[in] data Pointer to data to write.
- * \param[in] size Size of the data.
- */
-static inline void insert_into_buffer(buffer_t *buffer, const void *data, uint16_t size)
-{
-   uint16_t *msize = (uint16_t *)(buffer->data + buffer->wr_index);
-   (*msize) = htons(size);
-   memcpy((void *)(msize + 1), data, size);
-   buffer->wr_index += (size + sizeof(size));
 }
 
 /**
