+++ libtrap/src/ifc_tcpip_internal.h
@@ -41,6 +41,7 @@
  *
  */
 
+#include "ifc_socket_common.h"
 
 /** \addtogroup trap_ifc
  * @{
@@ -60,18 +61,6 @@
 #define DEFAULT_BUFFER_SIZE      100000    /**< Default buffer size [bytes] */
 #define DEFAULT_MAX_CLIENTS      20        /**< Default size of client array */
 #define DEFAULT_TIMEOUT_FLUSH    1000000   /**< Default timeout for autoflush [microseconds]*/
-
-#define BUFFER_COUNT_PARAM_LENGTH    13     /**< Used for parsing ifc params */
-#define BUFFER_SIZE_PARAM_LENGTH     12     /**< Used for parsing ifc params */
-#define MAX_CLIENTS_PARAM_LENGTH     12     /**< Used for parsing ifc params */
-
-typedef struct buffer_s {
-   uint32_t wr_index;                      /**< Pointer to first free byte in buffer */
-   uint64_t clients_bit_arr;               /**< Bit array of clients that have not yet received the buffer */
-
-   uint8_t *header;                        /**< Pointer to first byte in buffer */
-   uint8_t *data;                          /**< Pointer to first byte of buffer payload */
-} buffer_t;
 
 /**
  * \brief Structure for TCP/IP IFC client information.
