+++ libtrap/src/ifc_tls_internal.h
@@ -45,6 +45,8 @@
 #include <openssl/ssl.h>
 #include <openssl/err.h>
 
+#include "ifc_socket_common.h"
+
 
 /** \addtogroup trap_ifc
  * @{
@@ -64,18 +66,6 @@
 #define TLS_DEFAULT_BUFFER_SIZE      100000    /**< Default buffer size [bytes] */
 #define TLS_DEFAULT_MAX_CLIENTS      20        /**< Default size of client array */
 #define TLS_DEFAULT_TIMEOUT_FLUSH    1000000   /**< Default timeout for autoflush [microseconds]*/
-
-#define BUFFER_COUNT_PARAM_LENGTH    13     /**< Used for parsing ifc params */
-#define BUFFER_SIZE_PARAM_LENGTH     12     /**< Used for parsing ifc params */
-#define MAX_CLIENTS_PARAM_LENGTH     12     /**< Used for parsing ifc params */
-
-typedef struct tlsbuffer_s {
-    uint32_t wr_index;                      /**< Pointer to first free byte in buffer */
-    uint64_t clients_bit_arr;               /**< Bit array of clients that have not yet received the buffer */
-
-    uint8_t *header;                        /**< Pointer to first byte in buffer */
-    uint8_t *data;                          /**< Pointer to first byte of buffer payload */
-} tlsbuffer_t;
 
 /**
  * \brief Structure for TLS IFC client information.
@@ -125,7 +115,7 @@ typedef struct tls_sender_private_s {
     uint32_t buffer_size;                   /**< Buffer size [bytes] */
     uint32_t active_buffer;                 /**< Index of active buffer in 'buffers' array */
 
-    tlsbuffer_t *buffers;                   /**< Array of buffer structures */
+    buffer_t *buffers;                   /**< Array of buffer structures */
     tlsclient_t *clients;                   /**< Array of client structures */
 
     pthread_t accept_thr;                   /**< Pthread structure containing info about accept thread */
